Make test pass
