+++ skfda/misc/__init__.py
@@ -1,3 +1,4 @@
-from ._math import log, log2, log10, exp, sqrt, cumsum, inner_product
 from . import covariances, kernels, metrics
+from . import regularization
 from ._lfd import LinearDifferentialOperator
+from ._math import log, log2, log10, exp, sqrt, cumsum, inner_product
