+++ skfda/representation/basis/_bspline.py
@@ -218,122 +218,6 @@ class BSpline(Basis):
 
         return deriv_basis, np.array(deriv_coefs)[:, 0:deriv_basis.n_basis]
 
-    def _penalty(self, lfd):
-
-        coefs = lfd.constant_weights()
-        if coefs is None:
-            return NotImplemented
-
-        nonzero = np.flatnonzero(coefs)
-
-        # All derivatives above the order of the spline are effectively
-        # zero
-        nonzero = nonzero[nonzero < self.order]
-
-        if len(nonzero) == 0:
-            return np.zeros((self.n_basis, self.n_basis))
-
-        # We will only deal with one nonzero coefficient right now
-        if len(nonzero) != 1:
-            return NotImplemented
-
-        derivative_degree = nonzero[0]
-
-        if derivative_degree == self.order - 1:
-            # The derivative of the bsplines are constant in the intervals
-            # defined between knots
-            knots = np.array(self.knots)
-            mid_inter = (knots[1:] + knots[:-1]) / 2
-            constants = self.evaluate(mid_inter,
-                                      derivative=derivative_degree).T
-            knots_intervals = np.diff(self.knots)
-            # Integration of product of constants
-            return constants.T @ np.diag(knots_intervals) @ constants
-
-        # We only deal with the case without zero length intervals
-        # for now
-        if np.any(np.diff(self.knots) == 0):
-            return NotImplemented
-
-        # Compute exactly using the piecewise polynomial
-        # representation of splines
-
-        # Places m knots at the boundaries
-        knots = self._evaluation_knots()
-
-        # c is used the select which spline the function
-        # PPoly.from_spline below computes
-        c = np.zeros(len(knots))
-
-        # Initialise empty list to store the piecewise polynomials
-        ppoly_lst = []
-
-        no_0_intervals = np.where(np.diff(knots) > 0)[0]
-
-        # For each basis gets its piecewise polynomial representation
-        for i in range(self.n_basis):
-
-            # Write a 1 in c in the position of the spline
-            # transformed in each iteration
-            c[i] = 1
-
-            # Gets the piecewise polynomial representation and gets
-            # only the positions for no zero length intervals
-            # This polynomial are defined relatively to the knots
-            # meaning that the column i corresponds to the ith knot.
-            # Let the ith knot be a
-            # Then f(x) = pp(x - a)
-            pp = PPoly.from_spline((knots, c, self.order - 1))
-            pp_coefs = pp.c[:, no_0_intervals]
-
-            # We have the coefficients for each interval in coordinates
-            # (x - a), so we will need to subtract a when computing the
-            # definite integral
-            ppoly_lst.append(pp_coefs)
-            c[i] = 0
-
-        # Now for each pair of basis computes the inner product after
-        # applying the linear differential operator
-        penalty_matrix = np.zeros((self.n_basis, self.n_basis))
-        for interval in range(len(no_0_intervals)):
-            for i in range(self.n_basis):
-                poly_i = np.trim_zeros(ppoly_lst[i][:,
-                                                    interval], 'f')
-                if len(poly_i) <= derivative_degree:
-                    # if the order of the polynomial is lesser or
-                    # equal to the derivative the result of the
-                    # integral will be 0
-                    continue
-                # indefinite integral
-                derivative = polyder(poly_i, derivative_degree)
-                square = polymul(derivative, derivative)
-                integral = polyint(square)
-
-                # definite integral
-                penalty_matrix[i, i] += np.diff(polyval(
-                    integral, self.knots[interval: interval + 2]
-                    - self.knots[interval]))[0]
-
-                for j in range(i + 1, self.n_basis):
-                    poly_j = np.trim_zeros(ppoly_lst[j][:,
-                                                        interval], 'f')
-                    if len(poly_j) <= derivative_degree:
-                        # if the order of the polynomial is lesser
-                        # or equal to the derivative the result of
-                        # the integral will be 0
-                        continue
-                        # indefinite integral
-                    integral = polyint(
-                        polymul(polyder(poly_i, derivative_degree),
-                                polyder(poly_j, derivative_degree)))
-                    # definite integral
-                    penalty_matrix[i, j] += np.diff(polyval(
-                        integral, self.knots[interval: interval + 2]
-                        - self.knots[interval])
-                    )[0]
-                    penalty_matrix[j, i] = penalty_matrix[i, j]
-        return penalty_matrix
-
     def rescale(self, domain_range=None):
         r"""Return a copy of the basis with a new domain range, with the
             corresponding values rescaled to the new bounds.
