+++ skfda/representation/basis/_constant.py
@@ -38,20 +38,6 @@ class Constant(Basis):
         return (self.copy(), coefs.copy() if order == 0
                 else self.copy(), np.zeros(coefs.shape))
 
-    def _internal_representation(self):
-        return NumberRepresentation.from_basis(self)
-
-    def _penalty(self, lfd):
-        coefs = lfd.constant_weights()
-        if coefs is None:
-            return NotImplemented
-
-        internal_repr = self._internal_representation()
-
-        return np.array([[coefs[0] ** 2 *
-                          (self.domain_range[0][1] -
-                           self.domain_range[0][0])]])
-
     def basis_of_product(self, other):
         """Multiplication of a Constant Basis with other Basis"""
         if not _same_domain(self, other):
